+++ tests/_fixtures/fixture-compatibility-ledger.json
@@ -780,7 +780,8 @@
           "tests/_fixtures/_git/vue-termui",
           "tests/_fixtures/_git/vue-vben-admin",
           "tests/_fixtures/_git/vue-bits",
-          "tests/_fixtures/_git/shadcn-vue"
+          "tests/_fixtures/_git/shadcn-vue",
+          "tests/_fixtures/_git/splitpanes"
         ]
       },
       "evidence": {
